pyproject.toml
[project]
name = "bean-juice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.5.0",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.2",
    "scrapy>=2.14.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
